pyproject.toml
[project]
name = "Cocotst"
version = "0.1.1"
description = "Easily to code qqoffcial bot."
authors = [
    {name = "Sibuxiangx", email = "sibuxiang@proton.me"},
]
dependencies = [
    "uvicorn>=0.32.0",
    "fastapi>=0.115.4",
    "graia-broadcast>=0.23.5",
    "cryptography>=43.0.3",
    "loguru>=0.7.2",
    "requests>=2.32.3",
    "Launart>=0.8.2",
    "aiofiles>=24.1.0",
    "typer>=0.13.1",
    "richuru>=0.1.1",
    "watchdog[watchmedo]>=6.0.0",
    "pydantic>=2.0",
    "httpx>=0.28.1",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "black>=24.10.0",
    "isort>=5.13.2",
]
alc = [
    "arclet-alconna>=1.8.34",
    "arclet-alconna-tools>=0.7.10",
]
docs = [
    "mkdocs-material>=9.5.45",
    "mkdocstrings-python>=1.12.2",
    "mkdocs-gen-files>=0.5.0",
    "mkdocs-literate-nav>=0.6.1",
    "mkdocs-section-index>=0.3.9",
]
voice = [
    "graiax-silkcoder[ffmpeg]>=0.3.6",
]
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.black]
line-length = 110
target-version = ['py39']

[tool.pdm.build]
includes = ["src/cocotst"]